move command bot functionality to own bot base class, to avoid api change for future releases
